# api/docs/swagger.yaml
definitions:
...
  models.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Message:
    properties:
      receiverID:
        type: string
      value:
        type: string
    type: object
...
  models.RecoveryReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.SignupReq:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  title: Social server API
  version: "1.0"
paths:
  /auth/email/send:
    get:
      description: Send code to user verification/password recovery
      parameters:
      - description: user email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: SendCode
      tags:
      - Auth
  /auth/email/verify:
    get:
      description: User email code verification
      parameters:
      - description: verification type
        in: query
        name: type
        required: true
        type: string
      - description: code
        in: query
        name: code
        required: true
        type: string
      - description: user email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: VerifyCode
      tags:
      - Auth
  /auth/login:
    put:
      parameters:
      - description: user credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Login
      tags:
      - Auth
  /auth/recovery:
    put:
      description: Reset user password, requires verified email by VerifyCode
      parameters:
      - description: recovery request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RecoveryReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: PasswordRecovery
      tags:
      - Auth
  /auth/refresh:
    get:
      description: Update user's token pair
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: UpdateTokens
      tags:
      - Auth
  /auth/register:
    post:
      description: Register new user
      parameters:
      - description: user model
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.SignupReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register
      tags:
      - Auth
  /chat/history/{targetID}:
    get:
      description: Get chat history
      parameters:
      - description: user id
        in: path
        name: targetID
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GetHistory
      tags:
      - Chat
  /chat/message/{id}:
    delete:
      description: Delete message
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: DeleteMessage
      tags:
      - Chat
  /chat/ws/join:
    get:
      description: Join chat room via WebSocket
      parameters:
      - description: user id
        in: query
        name: id
        required: true
        type: string
      - description: message
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Message'
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Join
      tags:
      - Chat
  /files/update:
    post:
      description: Updates user's avatar
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: UpdateAvatar
      tags:
      - Files
  /files/upload:
    post:
      description: Upload file to cloud storage
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Upload
      tags:
      - Files
...
swagger: "2.0"